requirements.txt
# Core runtime dependencies (minimal, cross-platform)
pyserial==3.5
paho-mqtt==1.6.1
requests==2.32.5
pyModbusTCP==0.3.0
opcua==0.98.13
smbus2==0.4.3
protobuf==4.25.1
jsonschema==4.26.0
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.3
prometheus-client==0.24.1
elasticsearch==9.2.1
redis==5.0.1
psutil==7.2.1
jinja2==3.1.6

# Async protocol drivers (第四阶段优化)
aiomqtt==2.0.0
pymodbus==3.6.6
aiofiles==23.2.1

# Optional heavy/OS-dependent features are split:
# - requirements-optional-ai.txt (torch, etc.)
# - requirements-optional-capture.txt (pyshark + tshark binary)
# - requirements-optional-report.txt (weasyprint for PDF)
# - requirements-dev.txt (lint/test)

# Removed: fuzzfactory (no published wheel) – see docs for placeholder implementation.
kubernetes==28.1.0
Flask==2.3.3
fastapi==0.104.1
uvicorn==0.23.2
PyYAML==6.0.1
cryptography==41.0.5
pydantic==2.5.2
pyjwt==2.8.0
loguru==0.7.2
reportlab==4.0.7
weasyprint==60.1
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
pylint==3.0.2
